migrate availability/troubleshoot page
