+++ apps/web/app/future/availability/troubleshoot/page.tsx
@@ -1,0 +1,15 @@
+import { _generateMetadata } from "app/_utils";
+import { WithLayout } from "app/layoutHOC";
+
+import { getLayout } from "@calcom/features/troubleshooter/layout";
+
+import Troubleshoot from "~/availability/troubleshoot/troubleshoot-view";
+
+export const generateMetadata = async () => {
+  return await _generateMetadata(
+    (t) => t("troubleshoot"),
+    (t) => t("troubleshoot_availability")
+  );
+};
+
+export default WithLayout({ getLayout, Page: Troubleshoot });
